README.MD
# StudentDatabaseApp
This is a student database management system that takes an "n" number of students and enroll them in several courses of their choice and 
generates a unique student ID based on the student's current level/year of studies. 
After the enrolling, the system calculates the tuition fee based on the number of courses enrolled and the student is prompted to make payment. 
A receipt is generated for the students after successful enrollment. 
This is my first Kotlin project and I plan on improving it in future.